requirements.txt
moviepy==1.0.3
Pillow==10.3.0
requests==2.31.0
instagrapi==1.17.12
ffmpeg-python==0.2.0
google-cloud-aiplatform
